requirements.txt
Django
djangorestframework
markdown
django-filter
validate-docbr
requests